+++ include/dev/bcm2835_gpio.h
@@ -1,7 +1,7 @@
-#ifndef _AARCH64_GPIO_H_
-#define _AARCH64_GPIO_H_
+#ifndef _DEV_BCM2835_GPIO_H_
+#define _DEV_BCM2835_GPIO_H_
 
-#include <aarch64/bcm2835_gpioreg.h>
+#include <dev/bcm2835_gpioreg.h>
 
 struct resource_t;
 
@@ -14,4 +14,4 @@ void bcm2835_gpio_function_select(resource_t *r, unsigned pin,
 void bcm2835_gpio_set_pull(resource_t *r, unsigned pin,
                            bcm2838_gpio_gppud_t pud);
 
-#endif /* !_AARCH64_GPIO_H_ */
+#endif /* !_DEV_BCM2835_GPIO_H_ */
