+++ sys/drv/82371AB.c
@@ -6,8 +6,8 @@
 #include <sys/device.h>
 #include <sys/rman.h>
 #include <sys/bus.h>
-#include <mips/malta.h>
-#include <sys/pci.h>
+#include <dev/malta.h>
+#include <dev/pci.h>
 #include <sys/interrupt.h>
 #include <dev/piixreg.h>
 #include <dev/isareg.h>
