+++ sys/drv/bcm2835_rootdev.c
@@ -3,7 +3,7 @@
 #include <sys/bus.h>
 #include <sys/devclass.h>
 #include <aarch64/armreg.h>
-#include <aarch64/bcm2835reg.h>
+#include <dev/bcm2835reg.h>
 #include <aarch64/interrupt.h>
 #include <sys/kmem.h>
 #include <sys/pmap.h>
