+++ sys/drv/gt64120.c
@@ -29,14 +29,14 @@
 #define KL_LOG KL_DEV
 #include <sys/klog.h>
 #include <sys/mimiker.h>
-#include <mips/malta.h>
+#include <dev/malta.h>
 #include <mips/interrupt.h>
 #include <dev/i8259.h>
 #include <dev/piixreg.h>
 #include <dev/isareg.h>
 #include <dev/gt64120reg.h>
 #include <sys/interrupt.h>
-#include <sys/pci.h>
+#include <dev/pci.h>
 #include <sys/spinlock.h>
 #include <sys/libkern.h>
 #include <sys/bus.h>
